+++ apps/namadillo/src/App/Transfer/CustomAddressForm.tsx
@@ -1,25 +1,15 @@
 import { Input, Stack } from "@namada/components";
 
-import { Chain } from "@chain-registry/types";
 import { chain as osmosis } from "chain-registry/mainnet/osmosis";
-import { Dispatch, SetStateAction, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import namadaShieldedSvg from "./assets/namada-shielded.svg";
 import namadaTransparentSvg from "./assets/namada-transparent.svg";
-
-const bech32PrefixMap: Record<string, string> = {
-  osmo: "osmo",
-  znam: "namada",
-  tnam: "namada",
-  cosmos: "cosmos",
-};
 
 type CustomAddressFormProps = {
   onChangeAddress?: (address: string) => void;
   customAddress?: string;
   memo?: string;
   onChangeMemo?: (address: string) => void;
-  chain?: Chain;
-  setIsDisabled?: Dispatch<SetStateAction<boolean>>;
 };
 
 export const CustomAddressForm = ({
@@ -27,8 +17,6 @@ export const CustomAddressForm = ({
   onChangeAddress,
   memo,
   onChangeMemo,
-  setIsDisabled,
-  chain,
 }: CustomAddressFormProps): JSX.Element => {
   const [addressError, setAddressError] = useState<string | null>(null);
 
@@ -39,19 +27,6 @@ export const CustomAddressForm = ({
     return "";
   }, [customAddress]);
 
-  const getAddressPrefix = (address: string): string | undefined => {
-    return Object.keys(bech32PrefixMap).find((p) => address.startsWith(p));
-  };
-
-  const checkIfAddressMatchesChain = (
-    address: string,
-    chain: Chain | undefined
-  ): boolean => {
-    if (!chain || !address) return false;
-    const prefix = getAddressPrefix(address);
-    return prefix ? bech32PrefixMap[prefix] === chain.bech32_prefix : false;
-  };
-
   return (
     <Stack as="fieldset" gap={2}>
       {onChangeAddress && (
@@ -59,25 +34,7 @@ export const CustomAddressForm = ({
           <Input
             label="Recipient address"
             value={customAddress}
-            onChange={(e) => {
-              const enteredAddress = e.target.value;
-              onChangeAddress(enteredAddress);
-
-              let error: string | null = null;
-              let shouldDisable = false;
-
-              if (enteredAddress) {
-                if (!checkIfAddressMatchesChain(enteredAddress, chain)) {
-                  const prefix = chain?.bech32_prefix || "expected";
-                  error = `Only ${prefix} addresses are allowed`;
-                  shouldDisable = true;
-                }
-              }
-
-              setAddressError(error);
-              setIsDisabled?.(shouldDisable);
-            }}
-            error={!!addressError}
+            onChange={(e) => onChangeAddress(e.target.value)}
           >
             {iconUrl && (
               <i className="w-6 absolute right-4 top-1/2 -translate-y-1/2">
@@ -94,9 +51,7 @@ export const CustomAddressForm = ({
         <Input
           label="Memo"
           value={memo}
-          onChange={(e) => {
-            return onChangeMemo(e.target.value);
-          }}
+          onChange={(e) => onChangeMemo(e.target.value)}
           placeholder="Required for centralized exchanges"
         />
       )}
